+++ core/src/main/scala/ai/lum/odinson/lucene/search/OdinQueryWrapper.scala
@@ -6,7 +6,6 @@ import org.apache.lucene.index._
 import org.apache.lucene.search._
 import org.apache.lucene.search.spans._
 import ai.lum.odinson.lucene.search.spans._
-import ai.lum.odinson.state.State
 
 /** Wraps a SpanQuery to add OdinsonQuery functionality. */
 class OdinQueryWrapper(val query: SpanQuery) extends OdinsonQuery {
